FXML: Move Performance tests to tests/performance
